requirements.txt
streamlit == 1.37.1
langchain-community == 0.2.12
langchain-openai == 0.1.22
langchain-chroma == 0.1.3
python-docx == 1.1.2
openpyxl == 3.1.5
# protobuf==3.20.*
# pysqlite3-binary
